feat(review): 优化评论统计查询
